@@ -1,12 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 23
+        targetSdkVersion 27
         versionCode 1
         versionName "1.0"
     }
@@ -28,7 +27,7 @@ dependencies {
     testCompile 'junit:junit:4.12'
     testImplementation "org.robolectric:robolectric:3.3.2"
 
-    compile 'com.android.support:appcompat-v7:26.+'
-    compile 'com.nostra13.universalimageloader:universal-image-loader:1.9.4'
-    compile 'com.android.support:recyclerview-v7:26.+'
+    compile 'androidx.appcompat:appcompat:1.0.2'
+    compile 'com.nostra13.universalimageloader:universal-image-loader:1.9.5'
+    compile 'androidx.recyclerview:recyclerview:1.0.0'
 }
