@@ -1,7 +1,7 @@
 package com.xinlan.imageeditlibrary.editimage.fragment;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import com.xinlan.imageeditlibrary.editimage.EditImageActivity;
 
