@@ -1,7 +1,7 @@
 package com.xinlan.imageeditlibrary.editimage.view;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.util.AttributeSet;
 
 /**
